+++ R/function_PlotSubbasinRouting.R
@@ -331,11 +331,10 @@ PlotSubbasinRouting <- function(map, map.subid.column = 1, gd = NULL, bd = NULL,
 
     # Add various basemaps
     leafmap <- leafmap %>%
-      leaflet::addProviderTiles("CartoDB.Positron", group = "Map") %>%
+      leaflet::addProviderTiles("Esri.WorldGrayCanvas", group = "Map") %>%
       leaflet::addTiles(group = "Street") %>%
       leaflet::addProviderTiles("Esri.WorldTopoMap", group = "Topo") %>%
-      leaflet::addProviderTiles("Esri.WorldImagery", group = "Satellite") %>%
-      leaflet::addProviderTiles("CartoDB.PositronOnlyLabels", group = "Satellite")
+      leaflet::addProviderTiles("Esri.WorldImagery", group = "Satellite")
 
     # Save Image
     if (!file == "") {