+++ NEWS.md
@@ -12,6 +12,7 @@
 * Add CalculateRiverDistance() and LakeDataOutlets() functions
 * Add example [LakeData.txt](./inst/demo_model/LakeData.txt) file
 * Update examples for PlotSubbasinRouting() and fix bugs with line.weight and line.color
+* Replace CartoDB sources for Leaflet basemaps
 
 ## HYPEtools 1.6.7 (2026-04-08)
 *Enhancements and bug fixes*